+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to this project will be documented in this file, in reverse 
 ### Added
 
 - [#2](https://github.com/elie29/validator/issues/2) Add customizable messages on error.
+- [#3](https://github.com/elie29/validator/issues/3) Add ArrayRule to validator.
 
 ### Changed